home
-api chiamata get a endpoint feed per home
-inizializzato dinamicamente  showcase
-inizializzato dinamicamente continua a leggere
-inizializzato dinamicamente best of the week
-diversi fix
-conversioni di alcune classi da java a kotlin

+++ app/src/main/java/it/corelab/studios/airbooks/data/model/HOME/ItemReading.java
@@ -51,7 +51,7 @@ public class ItemReading {
     private Integer readings;
     @SerializedName("average_rating")
     @Expose
-    private Integer averageRating;
+    private Double averageRating;
     @SerializedName("lovers")
     @Expose
     private Integer lovers;
@@ -177,11 +177,11 @@ public class ItemReading {
         this.readings = readings;
     }
 
-    public Integer getAverageRating() {
+    public Double getAverageRating() {
         return averageRating;
     }
 
-    public void setAverageRating(Integer averageRating) {
+    public void setAverageRating(Double averageRating) {
         this.averageRating = averageRating;
     }
